chang int to size_t

+++ 2-selection_sort.c
@@ -29,7 +29,7 @@ void selection_sort(int *array, size_t size)
 		return;
 	for (i = 0; i < size - 1; i++)
 	{
-		int current_po = i;
+		size_t current_po = i;
 
 		for (j = i + 1; j < size; j++)
 			if (array[j] < array[current_po])
